refactor: split classes into separate files
